third_party/blink/web_tests/external/wpt/css/css-text-decor/text-decoration-serialization.tentative-expected.txt

This is a testharness.js-based test.
FAIL text-decoration shorthand serialization assert_equals: expected "underline rgb(0, 0, 0)" but got "underline solid rgb(0, 0, 0)"
Harness: the test ran to completion.
